+++ cuda/src/include/utils.h
@@ -20,7 +20,7 @@ inline double d_xorshift(uint32_t *state)
     return (double)x / (double)UINT32_MAX;
 }
 
-void load_parameters_from_file(char filename[], int *N_PART, int *BINS, float *DT, double *M, int *N_THREADS,
+void load_parameters_from_file(char filename[], int *N_PART, int *BINS, float *DT, float *M, int *N_THREADS,
                                unsigned int *Ntandas, int steps[], char inputFilename[], char saveFilename[],
                                bool *resume, bool *dump, float *sigmaL);
 
@@ -28,7 +28,7 @@ void read_data(char filename[], float *h_x, double *h_p, unsigned int *evolution
 
 void save_data(char filename[], float *h_x, double *h_p, unsigned int evolution, int N_PART);
 
-float energy_sum(double *d_p, int N_PART, unsigned int evolution, double M);
+float energy_sum(double *d_p, int N_PART, unsigned int evolution, float M);
 
 int make_hist(int *h_h, int *h_g, int *h_hg, int *d_h, int *d_g, int *d_hg, double *DxE, double *DpE, const char *filename, int BINS, float Et);
